로딩 화면 및 경험치 UI 기능 구현

+++ Assets/Scripts/Player/PlayerLevelSystem.cs
@@ -61,6 +61,6 @@ public class PlayerLevelSystem : MonoBehaviour
 
     public float GetExpPercent()
     {
-        return currentExp / requiredExp;
+        return (float)currentExp / requiredExp;
     }
 }
